BackdoorAccessFromTest/requirements.txt
# to run Appium tests from Robot framwork
robotframework
robotframework-appiumlibrary

# to connect to MQTT broker
paho-mqtt